final refactor commit
name change; deletion of old files
